Add a `DescriptorsHandlerMixin` class used throughout the http framework
