+++ proxy/http/server/plugin.py
@@ -15,13 +15,13 @@ from typing import Any, Dict, List, Optional, Tuple
 
 from ..websocket import WebsocketFrame
 from ..parser import HttpParser
+from ..descriptors import DescriptorsHandlerMixin
 
-from ...common.types import Readables, Writables, Descriptors
 from ...core.connection import TcpClientConnection
 from ...core.event import EventQueue
 
 
-class HttpWebServerBasePlugin(ABC):
+class HttpWebServerBasePlugin(DescriptorsHandlerMixin, ABC):
     """Web Server Plugin for routing of requests."""
 
     def __init__(
@@ -42,36 +42,6 @@ class HttpWebServerBasePlugin(ABC):
         Defaults to name of the class. This helps plugin developers to directly
         access a specific plugin by its name."""
         return self.__class__.__name__      # pragma: no cover
-
-    # TODO(abhinavsingh): get_descriptors, write_to_descriptors, read_from_descriptors
-    # can be placed into their own abstract class which can then be shared by
-    # HttpProxyBasePlugin, HttpWebServerBasePlugin and HttpProtocolHandlerPlugin class.
-    #
-    # Currently code has been shamelessly copied.  Also these methods are not
-    # marked as abstract to avoid breaking custom plugins written by users for
-    # previous versions of proxy.py
-    #
-    # Since 3.4.0
-    #
-    # @abstractmethod
-    def get_descriptors(self) -> Descriptors:
-        return [], []  # pragma: no cover
-
-    # @abstractmethod
-    def write_to_descriptors(self, w: Writables) -> bool:
-        """Implementations must now write/flush data over the socket.
-
-        Note that buffer management is in-build into the connection classes.
-        Hence implementations MUST call
-        :meth:`~proxy.core.connection.connection.TcpConnection.flush`
-        here, to send any buffered data over the socket.
-        """
-        return False  # pragma: no cover
-
-    # @abstractmethod
-    def read_from_descriptors(self, r: Readables) -> bool:
-        """Implementations must now read data over the socket."""
-        return False  # pragma: no cover
 
     @abstractmethod
     def routes(self) -> List[Tuple[int, str]]:
